got it working as it was before, hopefully
