+++ src/map/config/map.json
@@ -21,6 +21,28 @@
     7 
   ],
 
-  "defaultAnimationDuration": 300
+  "defaultAnimationDuration": 300,
+
+  "allColorsHex": [
+    "#FFB300",
+    "#803E75",
+    "#FF6800",
+    "#A6BDD7",
+    "#C10020",
+    "#CEA262",
+    "#007D34",
+    "#F6768E",
+    "#00538A",
+    "#FF7A5C",
+    "#53377A",
+    "#FF8E00",
+    "#B32851",
+    "#F4C800",
+    "#7F180D",
+    "#93AA00",
+    "#593315",
+    "#F13A13",
+    "#232C16" 
+  ]
 }
